[EE/BE][ET][Portable] Move scalar_to utils to scalar_utils.h
Differential Revision: [D75962656](https://our.internmc.facebook.com/intern/diff/D75962656/)

[ghstack-poisoned]
